Part 1-3 + Pard Drum + Layer 1 performance loading
Demo:
Very basic part performance load implementation.
Performance > Part 1 loads the performance into to TG 1-8 and adds 0 to each TG channel.
Performance > Part 2 loads the performance into to TG 9-16 and adds 3 to each TG channel.
Performance > Part 3 loads the performance into to TG 17-24 and adds 7 to each TG channel.
Performance > Part Drum loads the performance into to TG 25-32 and adds 8 to each TG channel.
Performance > Layer 1 loads the performance into to TG 33-40 and adds 0 to each TG channel.

This is just a test of how it sounds this way.
Every part has a reverb.

ToneGenerators=40 should be added to the minidexed.ini so you need rpi4 or rpi5.
It starts the Pi with fasts mode, so it is possible that you will need a heatsink.