Part 2 + Part 3 + Part 4 performance loading
Very basic part performance load implementation.
Performance > Part 2 loads the performance into to TG 9-16 and adds 3 to each TG channel.
Performance > Part 3 loads the performance into to TG 17-24 and adds 7 to each TG channel.
Performance > Part 4 loads the performance into to TG 25-32 and adds 11 to each TG channel.

This is just a test of how it sounds this way.
Every part has a reverb.

ToneGenerators=32 should be added to the minidexed.ini so you need rpi4 or rpi5.
It starts the Pi with fasts mode, so it is porssible that you will need a heatsink.